+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+
+0.2.1
+-----
+
+- Implemented `AwareLogger.isEnabledFor` to make more compatible as a drop-in
+  replacement for native loggers.
+
+
 0.1.0
 -----